Backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
python-multipart==0.0.26
tree-sitter==0.25.2
tree-sitter-python==0.25.0
tree-sitter-javascript==0.25.0
tree-sitter-typescript==0.23.2
